// src/app/main/_shared/models/query-movies/query-movies.interface.ts
export interface IQueryMovieRest {
  popularity: number;
  vote_count: number;
  video: boolean;
  poster_path: string;
  id: number;
  adult: boolean;
  backdrop_path: string;
  original_language: string;
  original_title: string;
  genre_ids: number[];
  title: string;
  vote_average: number;
  overview: string;
  release_date: string;
}

export interface IQueryMoviesRest {
  page: number;
  results: IQueryMovieRest[];
  total_pages: number;
  total_results: number;
}